Cleaned up SQS handling for downloads. Created a viable monitor/heartbeat process

+++ usaspending_api/download/management/commands/download_sqs_worker.py
@@ -1,11 +1,9 @@
 import botocore
-import functools
-import logging
 import os
 import signal
 import time
 
-from multiprocessing import Process, Value
+from multiprocessing import Process
 
 from django.conf import settings
 from django.core.management.base import BaseCommand
@@ -17,14 +15,12 @@ from usaspending_api.download.helpers import write_to_download_log as write_to_l
 from usaspending_api.download.lookups import JOB_STATUS_DICT
 from usaspending_api.download.models import DownloadJob
 
-CURRENT_DOWNLOAD_JOB = Value()
-CURRENT_MESSAGE = Value()
-DEFAULT_VISIBILITY_TIMEOUT = 60 * 30
-PROCESS_FAIL_COUNT = 0
-PROCESS_FAIL_MAX = 3
-PROCESS_RESTART_DELAY = 15
+CURRENT_MESSAGE = None
+DEFAULT_VISIBILITY_TIMEOUT = 60
+MONITOR_SLEEP_TIME = 5  # must be significantly less than DEFAULT_VISIBILITY_TIMEOUT
+LONG_POLL_SECONDS = 10
 
-logger = logging.getLogger("console")
+# logger = logging.getLogger("console")
 
 
 class Command(BaseCommand):
@@ -33,124 +29,56 @@ class Command(BaseCommand):
         for sig in [signal.SIGINT, signal.SIGQUIT, signal.SIGTERM]:
             signal.signal(sig, signal_handler)
 
-        download_service_app()
+        download_service_manager()
         logger.info("Bye")
 
 
 def signal_handler(signum, frame):
-    print("entered signal_handler({})".format(signum))
-    # import pdb; pdb.set_trace()
+    write_to_log({"message": "Received signal ({}). Releasing job to Queue".format(signum)})
+    set_sqs_message_visibility(CURRENT_MESSAGE, 0)
     raise SystemExit
 
 
 def download_service_manager():
-    download_app = create_and_start_new_process()
-
-    global PROCESS_FAIL_COUNT
-    global PROCESS_FAIL_MAX
+    global CURRENT_MESSAGE
 
     while True:
-        time.sleep(PROCESS_RESTART_DELAY / 2)
-        if download_app.exitcode not in (None, 0):
-            # Download job terminated with an error
-            raise SystemExit(1)
-        elif not download_app.is_alive():
-            # Download job completed with no error status codes (this is unexpected)
-            msg_prefix = "Download Service App unexpectedly ended with no errors."
+        CURRENT_MESSAGE = poll_queue(long_poll_time=LONG_POLL_SECONDS)
 
-            if PROCESS_FAIL_COUNT < PROCESS_FAIL_MAX:
-                PROCESS_FAIL_COUNT += 1
-                write_to_log(
-                    message="{} Creating new process in {}s".format(msg_prefix, PROCESS_RESTART_DELAY),
-                    download_job=CURRENT_DOWNLOAD_JOB,
-                )
-                time.sleep(PROCESS_RESTART_DELAY)
-                download_app = create_and_start_new_process()
-            else:
-                write_to_log(
-                    message="{} Exceeded {} retries. Check logs".format(msg_prefix, PROCESS_FAIL_MAX),
-                    download_job=CURRENT_DOWNLOAD_JOB,
-                    is_error=True,
-                )
-                raise SystemExit(1)
-
-
-def create_and_start_new_process():
-    download_app = Process(
-        name="Download Service Worker Proccess",
-        target=download_service_app,
-        args=(CURRENT_DOWNLOAD_JOB, CURRENT_MESSAGE),
-    )
-    download_app.start()
-
-    return download_app
-
-
-def error_wrapper(func):
-    @functools.wraps(func)
-    def wrap_for_errors(*args, **kwargs):
-        try:
-            download_job = kwargs.get('download_job') or args[0]
-            sqs_message = kwargs.get('sqs_message') or args[1]
-            value = func(*args, **kwargs)
-        except KeyboardInterrupt as e:
-            write_to_log(message="Process received external SIGNAL to terminate.", download_job=download_job)
-
-            if download_job:
-                download_job.error_message = None
-                download_job.job_status_id = JOB_STATUS_DICT["killed"]
-                download_job.save()
-            break
-
-        except FatalError as e:
-            write_to_log(message=str(e), download_job=download_job, is_error=True)
-            raise SystemExit(1)
-
-        except Exception as e:
-            logger.error(e)
-            write_to_log(message=str(e), download_job=download_job, is_error=True)
-
-            if download_job:
-                download_job.error_message = str(e)
-                download_job.job_status_id = JOB_STATUS_DICT["failed"]
-                download_job.save()
-        finally:
-            # Set visibility to 0 so that another attempt can be made to process in SQS immediately, instead of
-            # waiting for the timeout period to expire
-            if sqs_message:
-                set_sqs_message_visibility(sqs_message, 0)
-                write_to_log(message="Job released to queue", download_job=download_job)
-
-        return value
-    return wrap_for_errors
-
-
-@error_wrapper
-def download_service_app(download_job=None, sqs_message=None):
-    write_to_log("Started new Download Service App with pid {}".format(os.getpid()))
-    while True:
-        # second_attempt = True
-        sqs_message = poll_sqs_for_message(settings.BULK_DOWNLOAD_SQS_QUEUE_NAME, 3)
-
-        if not sqs_message:
+        if not CURRENT_MESSAGE:
             continue
 
-        write_to_log(message="Message Received: {}".format(sqs_message))
-        if sqs_message.body is not None:
-            # Retrieve and update the job
-            download_job = DownloadJob.objects.filter(download_job_id=int(sqs_message.body)).first()
-            if download_job is None:
-                raise FatalError("Download Job {} Missing in DB!!!".format(sqs_message.body))
+        download_job_id = int(CURRENT_MESSAGE.body)
 
-            # Retrieve the data and write to the CSV(s)
-            csv_generation.generate_csvs(download_job=download_job, sqs_message=sqs_message)
+        # If job, pass ID to new process
+        write_to_log(message="Message Received: {}".format(CURRENT_MESSAGE))
+        download_app = create_and_start_new_process(download_job_id)
 
-            # If successful, remove message from queue
-            if remove_message_from_sqs(sqs_message):
-                sqs_message = None
+        monitor_process = True
+        while monitor_process:
+            monitor_process = False
+            if download_app.exitcode == 0:  # If Process exits with 0, success. remove from queue
+                if remove_message_from_sqs(CURRENT_MESSAGE):
+                    CURRENT_MESSAGE = None
+            elif download_app.exitcode not in (None, 0):  # If process exits with error/raises exception, release message
+                update_download_record(download_job_id, "failed")
+            elif not download_app.is_alive():  # Process dies
+                update_download_record(download_job_id, "ready")
+                set_sqs_message_visibility(CURRENT_MESSAGE, 0)
+                time.sleep(MONITOR_SLEEP_TIME)  # allow the system to breathe before starting another job
+            else:
+                # Monitor process. Send heartbeats to SQS
+                set_sqs_message_visibility(CURRENT_MESSAGE, DEFAULT_VISIBILITY_TIMEOUT)
+                time.sleep(MONITOR_SLEEP_TIME)
+                monitor_process = True
 
 
-def poll_sqs_for_message(queue_name, wait_time=10):
+def poll_queue(long_poll_time):
+    # Future TODO: allow different types of queues (redis, etc)
+    return poll_sqs_for_message(settings.BULK_DOWNLOAD_SQS_QUEUE_NAME, long_poll_time)
+
+
+def poll_sqs_for_message(queue_name, wait_time):
     """ Returns 0 or 1 message from the queue"""
     try:
         queue = get_sqs_queue_resource(queue_name=queue_name)
@@ -161,24 +89,64 @@ def poll_sqs_for_message(queue_name, wait_time=10):
             MaxNumberOfMessages=1,
         )
     except botocore.exceptions.ClientError:
-        logger.exception("SQS connection issue. Investigate settings")
+        write_to_log(message="SQS connection issue. Investigate settings", is_error=True)
         raise SystemExit(1)
 
     return sqs_message[0] if sqs_message else None
 
 
+def create_and_start_new_process(download_job_id):
+    download_app = Process(
+        name="Download Service Worker Proccess",
+        target=download_service_app,
+        args=(download_job_id,),
+    )
+    download_app.start()
+
+    return download_app
+
+
+def download_service_app(download_job_id):
+    download_job = retrive_download_job_from_db(download_job_id)
+    write_to_log(message="Starting new Download Service App with pid {}".format(os.getpid()), download_job=download_job)
+
+    # Retrieve the data and write to the CSV(s)
+    csv_generation.generate_csvs(download_job=download_job)
+
+    return True
+
+
 def remove_message_from_sqs(message):
+    if message is None:
+        return
     try:
         message.delete()
-    except botocore.exceptions.ClientError:
-        logger.exception("Unable to delete SQS message. Investigate")
-        raise SystemExit(1)
+    # except botocore.exceptions.ClientError:
+    except Exception as e:
+        write_to_log(message="Unable to delete SQS message. Message might have been previously released or removed")
 
 
 def set_sqs_message_visibility(message, new_visibility):
+    if message is None:
+        return
     try:
         message.change_visibility(VisibilityTimeout=new_visibility)
-        # TODO: check existence instead of catching error
     except botocore.exceptions.ClientError as e:
-        logger.exception()
-        raise FatalError(e)
+        write_to_log(message="Unable to edit SQS message VisibilityTimeout. Might have already been released or removed")
+
+
+def retrive_download_job_from_db(download_job_id):
+    download_job = DownloadJob.objects.filter(download_job_id=download_job_id).first()
+    if download_job is None:
+        raise FatalError("Download Job {} record missing in DB!".format(download_job_id))
+    return download_job
+
+
+def update_download_record(download_job_id, status, message=None):
+    download_job = retrive_download_job_from_db(download_job_id)
+
+    download_job.job_status_id = JOB_STATUS_DICT[status]
+    if message:
+        download_job.error_message = str(message)
+
+    download_job.save()
